2007-05-08	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libnetworking/netdb.h: Change netent->n_net to uint32_t to comply
	with SUSv3.

+++ cpukit/libnetworking/netdb.h
@@ -105,7 +105,7 @@ struct	netent {
 	char		*n_name;	/* official name of net */
 	char		**n_aliases;	/* alias list */
 	int		n_addrtype;	/* net address type */
-	unsigned long	n_net;		/* network # */
+	uint32_t	n_net;		/* network # */
 };
 
 struct	servent {
